requirements.txt
torch
pandas
numpy
tqdm
pyyaml
python-chess
zstandard
scikit-learn
